+++ src/main/resources/db/migration/V008__reply.sql
@@ -1,0 +1,4 @@
+SET search_path TO samply;
+
+ALTER TABLE "reply" ADD COLUMN retrievedAt TIMESTAMP;
+ALTER TABLE "ntoken_query" DROP COLUMN selected_biobank;