+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [8.4.0 - 2021-08-25]
 ### Added
 - Change icinga project value by docker environment
+- Monitoring api for active checks
+### Changed
+- Share-dto version 5.1.0
+- Use gson while saving reply
 
 ## [8.3.0 - 2021-08-13]
 ### Added